requirements.txt
azure-functions
azure-eventhub 